feat: change to discordkt, split commands
