+++ commands.md
@@ -1,24 +1,32 @@
 # Commands
 
-## Key
-| Symbol     | Meaning                    |
-| ---------- | -------------------------- |
-| (Argument) | This argument is optional. |
+## Key 
+| Symbol      | Meaning                        |
+| ----------- | ------------------------------ |
+| (Argument)  | Argument is not required.      |
 
 ## Admin
-| Commands       | Arguments | Description                                   |
-| -------------- | --------- | --------------------------------------------- |
-| disable        | <none>    | Disable the bot                               |
-| enable         | <none>    | Enable the bot                                |
-| getPartySuffix | <none>    | Display current party mode suffix             |
-| setMode        | mode      | Set the mode to prefix or suffix              |
-| setOwner       | Member    | Set the owner & admin of the bot.             |
-| setPartySuffix | Suffix    | Set new party mode suffix                     |
-| setPrefix      | Prefix    | Set the bot's invocation prefix               |
-| setRole        | Role      | Set the role that will have symbols enforced. |
-| setSymbol      | Symbol    | Set the token to appear in nicknames.         |
-| toggle         | <none>    | Toggles the bot functionality                 |
-| toggleParty    | <none>    | Toggles party mode                            |
+| Commands | Arguments | Description                   |
+| -------- | --------- | ----------------------------- |
+| disable  |           | Disable the bot               |
+| enable   |           | Enable the bot                |
+| toggle   |           | Toggles the bot functionality |
+
+## Config
+| Commands  | Arguments | Description                                   |
+| --------- | --------- | --------------------------------------------- |
+| setMode   | mode      | Set the mode to prefix or suffix              |
+| setOwner  | Member    | Set the owner & admin of the bot.             |
+| setPrefix | Prefix    | Set the bot's invocation prefix               |
+| setRole   | Role      | Set the role that will have symbols enforced. |
+| setSymbol | Symbol    | Set the token to appear in nicknames.         |
+
+## Party
+| Commands       | Arguments | Description                       |
+| -------------- | --------- | --------------------------------- |
+| getPartySuffix |           | Display current party mode suffix |
+| setPartySuffix | Suffix    | Set new party mode suffix         |
+| toggleParty    |           | Toggles party mode                |
 
 ## Utility
 | Commands | Arguments | Description          |
