+++ src/main/kotlin/me/aberrantfox/hawk/botdata/Messages.kt
@@ -1,9 +1,8 @@
 package me.aberrantfox.hawk.botdata
 
-import me.aberrantfox.kjdautils.api.annotation.Data
+import me.jakejmattson.discordkt.api.dsl.data.Data
 
-@Data("data/messages.json")
 data class Messages(
     val USER_NICK_RESET: String = "Your nickname made it seem like you were a staff member. This has been automatically logged.",
     val STAFF_NICK_PROMPT: String = "You have updated your nickname and it does not contain the prefix. Here it is with the prefix: %1%"
-)
+): Data("data/messages.json")
