+++ src/main/resources/properties.json
@@ -1,6 +1,6 @@
 {
   "author": "${project.author}",
   "version": "${version}",
-  "kutils": "${kutils.version}",
+  "discordKt": "${discordKt.version}",
   "repository": "${project.repository}"
 }
